+++ packages/apps/shopify-app-remix/.eslintrc.js
@@ -1,4 +1,4 @@
-const config = require('../../.eslintrc');
+const config = require('../../../.eslintrc');
 
 module.exports = {
   ...config,