+++ packages/apps/shopify-app-remix/src/server/clients/admin/types.ts
@@ -150,7 +150,7 @@ export interface AdminApiContext<
    * Methods for interacting with the Shopify Admin GraphQL API
    *
    * {@link https://shopify.dev/docs/api/admin-graphql}
-   * {@link https://github.com/Shopify/shopify-api-js/blob/main/packages/shopify-api/docs/reference/clients/Graphql.md}
+   * {@link https://github.com/Shopify/shopify-app-js/blob/main/packages/apps/shopify-api/docs/reference/clients/Graphql.md}
    *
    * @example
    * <caption>Querying the GraphQL API.</caption>